+++ src/nr/types/utils/logging.py
@@ -19,48 +19,12 @@
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 # IN THE SOFTWARE.
 
-"""
-This module is used to design data models that can then be serialized and
-deserialized into/from JSON or YAML.
-"""
+class DeferredMessage(object):
 
-from .errors import *
-from .locator import *
-from .mixins import *
-from .types import *
-from .object import *
-from .collection import *
-from . import utils
+  def __init__(self, func, *args, **kwargs):
+    self.func = func
+    self.args = args
+    self.kwargs = kwargs
 
-
-def extract(value, py_type_def, locator=None, **options):
-  datatype = translate_field_type(py_type_def)
-  if locator is None:
-    locator = Locator.root(value, datatype, options)
-  else:
-    locator = locator.emplace(value, datatype, options)
-  return locator.extract()
-
-
-def store(value, py_type_def=None, locator=None, **options):
-  if py_type_def is None and isinstance(value, (Object, Collection)):
-    py_type_def = type(value)
-
-  datatype = translate_field_type(py_type_def)
-  if locator is None:
-    locator = Locator.root(value, datatype, options)
-  else:
-    locator = locator.emplace(value, datatype, options)
-
-  return locator.store()
-
-
-__all__ = (
-  errors.__all__ +
-  locator.__all__ +
-  mixins.__all__ +
-  types.__all__ +
-  object.__all__ +
-  collection.__all__ +
-  ['utils', 'extract', 'store']
-)
+  def __str__(self):
+    return str(self.func(*self.args, **self.kwargs))